+++ src/unicodedb.nim
@@ -6,11 +6,11 @@
 ## .. code-block:: nim
 ##   import unicode
 ##   import unicodedb
-##   echo category(Rune(0x860)) == "Lo"
-##   echo bidirectional(Rune(0x07F7)) == "ON"
-##   echo combining(Rune(0x860)) == 0
-##   echo name(Rune(32)) == "SPACE"
-##   echo lookupStrict("SPACE") == Rune(32)
+##   assert category(Rune(0x860)) == "Lo"
+##   assert bidirectional(Rune(0x07F7)) == "ON"
+##   assert combining(Rune(0x860)) == 0
+##   assert name(Rune(32)) == "SPACE"
+##   assert lookupStrict("SPACE") == Rune(32)
 ##
 ## There are more examples inluded
 ## within the tests module
@@ -22,7 +22,7 @@ import unicodedb/names
 
 export composition
 export decomposition, canonicalDecomposition
-export NfMasks, EProps, Props, properties,
+export NfMask, EProps, Props, properties,
        category, bidirectional,
        combining, quickCheck
 export name, lookupStrict