+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+v0.2.0
+==================
+
+* Add `contains` for `NfMask` usage,
+  so `nfcQcNo in quickCheck(x)` is supported
+* `NfMasks` was renamed to `NfMask`.
+  It's no longer `{.pure.}` and initial
+  letter of all properties are lowercase
+* Fix `lookupStrict` for nim 0.17.3
+
 v0.1.0
 ==================